Move classes out of legacy "silk.style.components" package
This is a substantial refactoring, because moving forward, we
are changing most usages of "ComponentStyle" to "CssStyle",
and narrowing the meaning of "component styling." Therefore,
it doesn't make sense to have a bunch of classes nested under
the "silk.components.style" subpackage anymore. Instead, many of
classes now simply live under the "silk" package directly
(i.e. "silk.animation", "silk.breakpoints", and "silk.style")

In this new world, instead of trying to have one 'ComponentStyle'
class act as both the new approach and old approach
simultaneously (i.e. one where the name comes from the Kotlin
code itself and another where the name comes from KSP / a
@CssName annotation), we now have two copies of the class:

1) The previous one, now deprecated ("silk.components.style.ComponentStyle")
2) The new one ("silk.style.component.ComponentStyle")

KSP processing code and the SilkTheme API have both been extended to
allow support for handling both versions, so we won't break on legacy
codebases, even as we push people to the new APIs.

In addition to "ComponentStyle", "ComponentVariant" and "Keyframes"
classes have also been forked, so we are currently maintaining a
legacy version and a new version of them as well.