Change selector name we register into the stylesheet back to ".base.variant"
... back from just ".variant"

This ensures two things...

1) That the specificity of the variant will always be higher
   than the specificity of the base style (and will thus get
   layered on top)
2) That a variant's styles will not be applied on an element
   that does not also apply the base styles as well

While technically we don't need to do this, as we always add
variants into a stylesheet AFTER all base styles (so its
specificity will be higher), and Kobweb code guarantees that
applying the base modifier + variant will include both class
names on the final element, it's still a nice thing to do in
order to express clarity of intention.

We also delete some related obsolete code in ComponentVariant
in order to simplify the codebase.