Rename "StyleModifiers" to "StyleScope"
This new name makes more sense after a previous refactoring
which introduced the names `CssStyleScope` and `CssStyleBaseScope`

The only wrinkle is that Compose HTML provides its own "StyleScope"
class, but in practice users never import it directly so the conflict
is not expected to cause issues. In fact, it probably emphasizes that
this is a good rename, since both our `StyleScope` and theirs
represent the same concept.