Add KSP default prefix support
After this change, a library author can write this in
their build script:

```
kobweb {
  library {
    cssPrefix.set("bs") // for "Bootstrap"
  }
}
```

at which point, any CssStyle, ComponentStyle, ComponentVariant, or
Keyframes block that the project declares will prepend the prefix
in front of the CSS name generated for it.

For example, `val Button = CssStyle { ... }` will generated the
CSS name "bs-button" and not just "button", given the prefix above.

The prefix you choose is not too crucial, as users won't see it unless
they inspect the DOM tree. However, as a library author, this is a best
practice, in order to reduce the chance of your style names and styles
defined in the user's code from colliding.