Tighten up logic that checks for CssStyle<ComponentKind> being in the same file
The previous logic, a little sloppy, was crashing on this line:

```
val TestSize = ButtonSize(...)
```

because we were trying to extract a generic parameter from TestSize
(thinking it was a CssStyle<...>) when it didn't have one.

Now, we create a KStype for `CssStyle<out ComponentKind>` and we
use that new type to check things more robustly. For example, in
the above case, we now can skip over running logic against
`TestSize` because we know that `TestSize` is not a
`CssStyle<ComponentKind>`