Add support for extending CssStyles
This looks like:

```
val BaseStyle = CssStyle { ... }
val ExtendedStyle = BaseStyle.extended { ... }
```

With this declared, doing the following:

```
Box(ExtendedStyle.toModifier())
```

will ensure that styles from the base and extended styles
are applied.

This may seem similar to defining variants for
ComponentKind styles, except this doesn't enforce type
safety, which variants are good for. Also, this is more a
case where the user will interact with the extended CSS
style directly, vs. variants where the user will interact
with the base style and then specify zero or more variants
on top of it.