Initial support for CSS cascading layers
Later, we may expand on this; but at the very least, this
ensures that any style a user defines will take precedence over
our own. This means a user will never define a style in their
own code and seethe because it's not getting applied --
although of course they will have to be careful that they don't
accidentally break some style that Silk depends on.

Bug #524