Break up "framework" CSS layer
We now have
- "component-styles"
- "component-variants"
- "restricted-styles"
- "unspecified-styles"

(in that order of precedence)

which get equally applied to styles defined by the
Silk framework as well as user styles. This should
be a more consistent approach moving forward.

Users can still pass in custom layer names, but these
will have the effect of overriding the default layer
their style would have been applied to, instead of
setting a layer on something without layers.