KSP: Clear the "consumedComponentKinds" set after each file
This is not in response to an issue we actually faced, but our intention
was always to only scope the set's values to a single file. Originally,
I assumed a new visitor was created per file, or that a visitor could only
run on a file once; but with things like multiple KSP rounds, this might
not be true! So now we clear the set, just to be extra safe.