Fix markdown logic so that MarkdownContext is ALWAYS provided
The previous logic would only provide a context if some frontmatter
was found, but that was not intentional. If you look at the docs, the
idea is supposed to be that ANY page generated from markdown gets
markdown context.

Also, MarkdownContext includes information about the markdown
path it came from, which is generally useful regardless of whether
frontmatter was found in it or not.

By updating this text rendering logic to always happen, we could
simplify code elsewhere that was duplicating some of the logic.