Put "PackageMapping" logic back into ConvertMarkdownTask
If a user puts a markdown file under
"public/resources/markdown/hello_there/123/test.md", then
the final route should respect that and be
"/hello_there/123/test"!

A while ago, we used to have logic that added package mappings when
converting markdown to Kotlin in order to handle this, but when we
refactored the way we generated routes to use
hyphens (in commit de88d625b), we removed it. It turns out, this was
premature.

This commit adds logic back in, slightly modified to take the current
design of Kobweb into account.