Fix issue where scrolling to fragment on the same page got broken
Also, added some long pages that interact with each other to make it easier
to test / reproduce this issue.

Before this fix, if you went to either of the long pages and clicked on the first
link on either page (the "go halfway down" link), it wouldn't do anything, because
what happened in that case was we waited for a mutation event to take place but it
never would because the page was already loaded.

Now, we go through a different path if we haven't changed pages, searching for the
fragment in the existing document instead of waiting for it to get created.