Detect cases where a user might have prefixed the route prefix an extra time
In the previous commit, we started adding support so that
`ctx.router.tryRoutingTo` and `ctx.router.navigateTo` now auto-prepend a
site prefix, if present. This is more intuitive behavior I've come to
believe, but this can break old sites that had to work around this using
code like `ctx.router.tryRoutingTo(RoutePrefix.prepend("/"))`.

That code can now be simplified to `ctx.router.tryRoutingTo("/")`, but
old codebases will continue to work; they'll simply report a warning in
the console when that path is detected.