Remove autoPrefix approach
We now ALWAYS prepend the base path in our routing code.
This makes the logic much simpler to understand, and in
general, users shouldn't need to think about this stuff.
It also seems to be the behavior of other web frameworks.

In practice, I don't think any user was using `autoPrefix`
except for us, internally, for a handful of special cases
(anchors and setting the URL directly from the window
browser bar). At this point, we've tweaked the code to
handle those issues in a more direct way (passing full paths
including the site origin into `navigateTo`)