Temporarily remove the deprecation warning for routePrefix
Unfortunately, the Kobweb CLI does not yet know about the new basePath
parameter -- it will need to update its library dependency on Kobweb
to learn about it.

So until we set that up and push out a new CLI, we probably shouldn't
push people to update to "basePath" yet, because then the older
CLIs don't support it.