Update base path section in README with BasePath.prepend example
I think the original documentation wasn't clear about real cases where
you'd have to use BasePath.prepend, because a lot of Compose HTML
widgets will never be base-path aware.

I also removed a long technical note because, in retrospect, I don't
think it helped that much but it added a bunch of noise to a README
that is already long enough.