Add initial implementation for WebGL2RenderingContext
... and CanvasGl2 that uses it.

This additional probably requires a closer look, but from the
MDN docs, WebGL2RenderingContext implements WebGLRenderingContext,
so for now, just basically duplicating it seems like it should be
a reasonable start. Eventually, we should add check for additional
methods that should be present in GL2 that's not in the base class.