Fix error with URLStreamHandlerFactory being set multiple times.
This was causing the server to fail to live reload, because `setURLStreamHandlerFactory`
is really strict about only one being set per JVM run. I'm not sure how this wasn't
crashing before -- maybe something changed in a more recent JDK, or I missed this
during a previous refactoring.

To reproduce, I simply modified an @Api method while the server is running and it
crashed.