Bound RouteTree<T>'s type to Any
A RouteTree's data type should NEVER be null. Internally, its
algorithm distinguishes between null data and non-null data to
make important decisions about whether a route was previously
registered or not, so someone declaring a `RouteTree<Data?>`
doesn't even make sense, and it was a mistake not to constraint
the bound.