Introduce `@Layout` annotation
With this commit, you can tag a `@Page` with an additional `@Layout`
annotation which identifies a parent layout for that page.

A layout method must take a `content` lambda parameter and, optionally,
a `PageContext` parameter:

```kotlin
@Composable
fun ExampleLayout(ctx: PageContext, content: @Composabl () -> Unit) {
  ...
}

@Layout(".components.layouts.ExampleLayout")
fun SomePage(ctx: PageContext) {
  LaunchedEffect(Unit) {
    ctx.data["title"] = "Some Page"
  }
}
```

In followup commits, we will allow supporting layouts in markdown and
for tagging packages. But for this commit, we only support the annotations
in code on pages (and layouts)
```

In followup commits, we will allow supporting layouts in markdown and
for tagging packages. But for this commit, we only support the annotations
in code on pages (and layouts).