In tryRoutingTo, download external content in place...
... instead of in a new tab.

This should prevent a temporary new tab from opening up and
disappearing every time you click on some file that should be
served directly from the server.

Also, in the past, you could hit a bug which we think we've since
fixed, but when triggered, you could have a browser spawning
infinite new tabs. With this logic, we would have nuked the current
tab... still not great, but at least it would have been contained.