Downgrade "No root route found" from an error into a warning
Before, when exporting a site that had no root "Index.kt" page, we
would output this:

```
No root route was defined for your site. This means if people visit
your website URL, they'll get a 404 error. Create a `@Page` in a root
`pages/Index.kt` file to make this warning go away.
```

I mean, I even wrote at the time it was a warning, so it was clear that
the original intention was not to block the export.

At this point, it has been converted into a genuine warning AND
we've added a build script property that will allow users to suppress it.