Rename root element ID from "root" to "_kobweb-root"
This gives us less chance for a user to accidentally name their own
element the same thing. This actually happened in our docs site, where
we created a section for an article and called it "root" but when the
user clicked on it and tried to link to "some-article#root", it scrolled
to the top, trying to focus our root element instead.