Make various interceptUrls block properties internal
... so we go from suggesting that people should using
methods to configuring the block to requiring them.

The methods basically feel better to read:

```
// Before
rejects.add("...")

// After
reject("...")
```

and, in the case of self-hosting, if excludes are provided,
have an experience that is both more ergonomic and also
cannot leave users in an unexpected invalid state.

```
// Before
selfHosting {
   enabled.set(true)
   excludes.add("...")
}

// After
enableSelfHosting(excludes = setOf("..."))

// This is no longer possible:
selfHosting {
  excludes.add("...")
  // Uh oh, self hosting still not enabled!
}
```