Add reference to conf.yaml title in Kobweb backend
This was the only field that wasn't technically used by the
backend before, which made it stand out as an oddity. Now
it is not!

It can be argued that the title does not HAVE to be referenced
by the serve logs, of course, and we can always pull it out if
we decide some day to remove the title from the conf.yaml file
and make it purely a frontend only concept.

However, it seems pretty reasonable to declare the title inside
the conf.yaml file, so might as well lean into it?