Make dynamicParams available for querying
Maybe no one will ever need it, but I realize if someone wants to
know if they're on a dynamic route, there's really no to test that
right now. But after this change, a user could add
`ctx.route.dynamicParams.isNotEmpty()`. So it seems like maybe we
should go aheand and expose it.