Make linking between markdown files more robust
Before, logic assumed that there would ever only be a single
markdown root and that it always represented the root of all
pages.

For example, "markdown/a/b/c/Test.md" would correspond to the route
"/a/b/c/test"

But now, markdown roots can be a bit more flexible. Specifically,
you can change it so that "markdown/pages/a/b/c/Test.md" should
correspond to "/a/b/c/test", but the old logic would include the
prefix "/pages" in that too.