Fix logic for generating route from markdown path
Before, if we had a file in a path that couldn't convert to
a package directly (e.g. 2025/03/27), we would generate a
best-fit package (e.g. "_2025._03._27") and then set the
output path to that to match (so far so good).

Unfortunately, we then based our route override value on
path, which ended up using the transformed (and wrong)
segment values.

Now, we base our route on the incoming source file path,
before it got transformed into package friendly values.