Rename "DisconnectedStream" to "LimitedStream"
Imagine writing a function that could take a parameter which
could accept an active stream as well as a disconnected one.

```
process(stream: DisconnectedStream)
```

That is kind of hard to read; you wouldn't think about passing in
an active stream into that method, surely.

So we rename it here to "LimitedStream", which captures the same
idea but doesn't have such an opinionated declaration about the
stream's state.