Every stream connection now gets its own unique ID
Before, we had a unique ID per client, but I was recently
working on a demo where I wanted to create two different
ApiStream instances on the same client connecting to the
same endpoint, and this wouldn't work -- both streams got
data that should have only gone to the other.

In 99.99% of all cases, most people will only ever create
a single `ApiStream` per client, but as a mental model,
there was no reason why they shouldn't have been able to
create two or more. At the very least, we allowed it, so
now the behavior should work.