Fix issue with init route / layout ordering
After this fix, we now ensure that both init route calls
and layouts are done in the right order. Specifically,
if you have a child page, parent layout, and grandparent layout...

1. init route - child
2. init route - parent
3. init route - grandparent
4. render - grandparent
5. render - parent
6. render - child

Before, the init route order was backwards, where it would
initialize the child page first (good!) but then run through
ancestors in the opposite order.