Deprecate support for `blend` to `Background.of`
It was a neat idea but in practice it is probably a bit too
clever. (When blend mode is set, it would additionally and automatically set the "background-blend-mode" property too.)

However, this could break in very unexpected ways. For example,
if someone created a background value in a style variable
with the blend mode set on it, that value would be lost when
the CSS variable was assigned.