Move scroll logic from navigate phase to render phase
Previously, we would perform scrolls directly inside the `navigateTo` /
`tryRoutingTo` methods, but this would cause us to occasionally try to
scroll too quickly, before the page was ready.

Now we set a callback which is triggered automatically when a render
happens -- and at that point, the scroll logic is now much more in
sync with the actual page contents.

We detected this issue originally on the Kobweb site, where you could
scroll down a long page, open up the docsearch modal, and visit another
page. Kobweb is SUPPOSED to set the page to the top at that point, but
what was happening is because the scroll happened so quickly, it would
get overriden by the docsearch modal closing on the previous page (which,
as a side effect, would also try to set the local page's scroll position).
Now that we wait until the new page starts rendering first, this issue
no longer occurs.