Fix SVG attr methods not autocompleting
Looks like an issue with autocomplete got broken in
K2, which unfortunately makes working with SVGs very
difficult right now. So we feel pressure to make these
interfaces public again for now.

To avoid totally random pollution of the user's
codespace, we prefix the interfaces now with Svg-

We reported an issue to YouTrack:
 https://youtrack.jetbrains.com/issue/KT-77311

and we can revert these back to internal if / when
they fix it.