Fix live reloading not working when multiple sites exist in the project
The `registerIfAbsent` callback runs only once for the build, meaning that the project in which it runs is not necessarily the project that will be executed. As a result, previously only the first configured project would receive task completion events, while now events are sent to the specific project being executed.