Change all Yaml deserialization settings to non strict mode
This will allow us to be able to add new fields to our APIs
without worrying about breaking older CLIs out in the wild
that haven't been upgraded.

Unfortunately, it will probably take a long time for this
change to propagate through to most users, but better to
start sooner than later.