Router: treat //example.com as an external link
A link like `<a href="//example.com">` means to open the target in the
current procol scheme, such as http or https.

Until this change the Router and Link component treated such a target as
an internal link, dispatching to the current origin rather than the
external domain.

Example: linking to //example.com/page from
https://example.org used to result in
https://example.org/example.com/page

With this change it will result in https://example.com/page