Fix quoted strings issue in Markdown parsing
We revamped yaml parsing a little while back and it appears we
broke quotes within strings at that point.

For example, with:

```yaml
test1: "This is a string with quotes in it," he pointed out.
test2: "String fully wrapped in quotes"
```

test1 returns `"This is a string with quotes in it," he pointed out.` and test2 returns `String fully wrapped in quotes`

When we then go forward and add these strings into our code, test1
will break because of the introduced quote:

```kotlin
// generated code
addScalar("test1", ""This is a string with quotes in it," he pointed out.")
```

So we need to escape those quotes. Ironically, there was code already there
UNESCAPING those quotes, which either means this was just a dumb
logic error, or our previous yaml parsing approach was doing something
different.